Added new test classes and refactored a bit for tests.
